requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tifffile==2023.9.26
torch>=2.0.1+cu118
torchvision>=0.15.2+cu118
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
